Move global member Twig variables to its own extension, using the SecurityBundle TokenStorage
